megano/requirements.txt
amqp==5.1.1
arrow==1.2.3
asgiref==3.6.0
async-timeout==4.0.2
billiard==4.1.0
blessed==1.20.0
celery==5.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
diploma-frontend @ file:///mnt/c/Users/Victor/PycharmProjects/python_django_diploma/diploma-frontend/dist/diploma-frontend-0.6.tar.gz
Django==4.2.1
django-db-queue==3.0.0
django-dramatiq==0.11.2
django-filter==23.2
django-picklefield==3.1
django-q==1.3.9
django-rq==2.8.1
djangorestframework==3.14.0
dramatiq==1.14.2
dramatiq-dashboard==0.4.0
gevent==22.10.2
greenlet==2.0.2
Jinja2==3.1.2
kombu==5.3.0
MarkupSafe==2.1.3
pika==1.3.2
Pillow==9.5.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
rq==1.15.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
watchdog==3.0.0
watchdog-gevent==0.1.1
wcwidth==0.2.6
zope.event==4.6
zope.interface==6.0
